Move base cell_object to dedicated file
